+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# WIP 0.11.12
+
+* Add `labels` arg and property to `ComputeStep` base class
+* Add `labels` arg to `BatchTransform` and `BatchTransformStep`
+* Add `labels` arg to `BatchGenerate` and `DatatableTransformStep`
+
 # 0.11.11
 
 * Add formatting through `traceback_with_variables` in `event_logger`